requirements.txt
Click==7.0
Flask==1.0.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
httplib2==0.12.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauth2client==4.1.3
python-dotenv==0.10.2
requests==2.22.0
sqlalchemy==1.3.3
Werkzeug==0.15.4
